refactor: The code logic of DatProjectFactory is in DatProjectUtil
